Pass DB env vars to subprocess
Why:

The user's script may use kleio's logger to connect to the database. If
the configuration of the database used with kleio is not from global
configuration files, then the instantiation of the database within the
user script will be incoherent since it can only use global
configuration.

How:

Set env vars for DB configuration inside the wrapper and pass them to
the user's script variables. This way we are sure the database within
the script will be the same because instantiation of kleio's database
within the user script can only use global config or environment
variables. Not that environment variables have precedence over global
configuration.