Add better script detection
Why:

fetch_user_script was currently only looking for python in first
position of commandline to pick the next argument as the script,
otherwise it would pick the first argument as the script. This obviously
fails in many situations.

How:

Use regex to pick first occurrence of python* {options} {script-name}.py.

This only works if meaningful python script is the first occurrence.