Change db.write_file to use files on local FS
Why:

Massive check-pointing of models causes the DB to fill rather quickly.
A temporary solution while a proper database setup is not available is
to use the local file system. That means trials may *NOT* be resumable
from different hosts, since the files cannot be found.

How:

KLEIO_DATABASE_FILE_DIR provides a path where to save artifacts.
Associated metadata is saved in database and the id serve as the file
name.