Pass parent endtime when cli run auto branches
Why:

It was previously branching from the beginning of the parent trial.
While that make sense for cli `save`, because the use wants to start
another identical cli execution from the start, it does not make sense
for cli `run`. If it finds a trial to resume and need to branch because
of host of version changes, it should not point to the beginning of the
parent and start from the beginning, it should point to end of parent
and resume from there.