Try to fix resuming bugs with asyncio
Why:

There is an error because of loops being closed before some coroutine
ends. Not sure why precisely that happens, probably because of the
update_task which is started independently and canceled manually. The
others do not seem to follow.

How:

Use the solution cancel, loop_until_complete (cancellation) than close
loop.

It seems to work, but now if we try to suspend remotely by changing the
status with cli command suspend, the update_task coroutine exits but the
other coroutine continues... Will need to fix this as well.