Set logger's insert stats tags to local ones only
Why:

Tags should be soon integrated as part of the immutable constituents of
a trial, making it possible to run multiple times the same execution
with different tags. Because of this, there is no point in inserting
current trial's tags in its own statistics. Only the tags of the trial
inserting the statistics should be saved. To make it clear, when there is a trial A
inserting statistics in a trial B, only the tags of A are saved along
with the statistics.