Add support for AnalysisLogger to BackupLogger
Why:

Process which uses AnalysisLogger cannot be executed without kleio if we
do not provide support a backup logger. In the case of backup logger
however, no writing rights are provided thus the BackupLogger will
give read access to the db but will only print what is attempted to be
logged, just like the BackupLogger does for Logger.

How:

Load a view of the trial if BackupLogger is built as an AnalysisLogger.
All methods for reading are as AnalysisLogger, but writing methods only
prints the data.