Add debugging print...
Why:

This bug this is very difficult to reproduce. Sometimes when branching
the parent trial has no status and trying to access the end_time raises
an error. There problem is that the parent actually has a status! But
for some reason at run time we end up with a parent without any status.
It seems not to be because of the interval in refers based on current
debugging. Must probable cause seems to be ID mismatch at creation time,
with configuration leading to different hash because of bad sorteddict.