Fix weird property/getattr bug
Why:

When building children of a TrialNode there was some weird behavior,
where __getattr__ would be called on TrialNode objects for item, parent
or children attributes whereas it should only be called on non-existing
attributes. Possibly this is because there are properties rather than
standard attributes. Anyway, a simple fix is to try again to get
attribute and on fail let it go to the item.