Set tags properly on auto-branching
Why:

When running with run --tags 'some-tags', the process found is only
tagged with the tags in command-line when auto-branching. A branched
children should have the same tags as the parent, not a subset of it.
Note that the parent MUST contain the tags passed in commandline, this
is how we find it, but it can also contain additional tags. It's those
tags that was previously missing on the children.