Set indexes in attributes and for trial.report
Why:

The indexes were not set for all attributes because the flag was set on
the class, but different objects can be built for different collections
in the database. There should be a flag for each collection.

How:

Turn the boolean flag into a set that contains strings of each
collections for which an index has been built.