Set attributes ids based on last one
Why:

If there is any error it is painful to fix the DB when the IDs are based
on the number of elements. If we simply used the last id, then it is
just as robust to race-conditions and can easily adapt if there is an
error and a gap is created in the index.